+++ src/cli/commands/public-cmds/log-cmd.ts
@@ -8,12 +8,13 @@ export default class Log extends Command {
   name = 'log <id>';
   description = `show components(s) tag history.\n  https://${BASE_DOCS_DOMAIN}/docs/view#log`;
   alias = '';
-  opts = [];
+  // @ts-ignore
+  opts = [['r', 'remote', 'show log of a remote component']];
   migration = true;
   remoteOp = true; // should support log against remote
 
-  action([id]: [string]): Promise<any> {
-    return getComponentLogs(id).then(logs => {
+  action([id]: [string], { remote = false }: { remote: boolean }): Promise<any> {
+    return getComponentLogs(id, remote).then(logs => {
       Object.keys(logs).forEach(key => (logs[key].tag = key));
       return R.reverse(R.values(logs)).map(
         R.evolve({